% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getClustFunParams}
\alias{getClustFunParams}
\title{getClustFunParams}
\usage{
getClustFunParams()
}
\value{
cluster params
}
\description{
get cluster function parameters
}
\note{
\code{getClustFunParams}
}
\examples{
    x <- getClustFunParams()
}
