% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getCustomColors}
\alias{getCustomColors}
\title{getCustomColors}
\usage{
getCustomColors(num = 1)
}
\arguments{
\item{num, }{num for repetions of the controls}
}
\value{
color range
}
\description{
get Custom Colors
}
\note{
\code{getColRng}
}
\examples{
    x <- getCustomColors()
}
