% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getDistFunParams}
\alias{getDistFunParams}
\title{getDistFunParams}
\usage{
getDistFunParams()
}
\value{
pals
}
\description{
get distance metric parameters
}
\note{
\code{getDistFunParams}
}
\examples{
    x <- getDistFunParams()
}
