% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getHeatmapControls}
\alias{getHeatmapControls}
\title{getHeatmapControls}
\usage{
getHeatmapControls()
}
\value{
HeatmapControls
}
\description{
Generates the left menu to be used for heatmap plots
}
\note{
\code{getHeatmapControls}
}
\examples{
    x <- getHeatmapControls()
}
