% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getMenuIcon}
\alias{getMenuIcon}
\title{getMenuIcon}
\usage{
getMenuIcon()
}
\value{
returns the left menu according to the selected tab;
}
\description{
Returns the menu icons
}
\note{
\code{getMenuIcon}
}
\examples{
    x <- getMenuIcon()
}
