% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getPal}
\alias{getPal}
\title{getPal}
\usage{
getPal(num = 1)
}
\arguments{
\item{num, }{num for repetions of the controls}
}
\value{
pals
}
\description{
get pallete
}
\note{
\code{getPal}
}
\examples{
    x <- getPal()
}
