% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getQCReplot}
\alias{getQCReplot}
\title{getQCReplot}
\usage{
getQCReplot(cols = NULL, conds = NULL, datasetInput = NULL,
  input = NULL)
}
\arguments{
\item{cols, }{the dataset to use}

\item{conds, }{the dataset to use}

\item{datasetInput, }{the dataset to use}

\item{input, }{user input}
}
\value{
the panel for QC plots
}
\description{
Prepares QCplots for comparisons and others
}
\note{
\code{getQCReplot}
}
\examples{
    x <- getQCReplot()

}
