% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getSizeControls}
\alias{getSizeControls}
\title{getSizeControls}
\usage{
getSizeControls(name = "sizecontrol", h = 500, w = 640)
}
\arguments{
\item{name, }{name of the size controls}

\item{h, }{height}

\item{w, }{width}
}
\value{
size controls
}
\description{
Generates the size controls
}
\note{
\code{getSizeControls}
}
\examples{
    x <- getSizeControls()
}
