% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getTableModal}
\alias{getTableModal}
\title{getKEGGModal
prepares a helpbutton for to go to a specific site in the documentation}
\usage{
getTableModal()
}
\value{
the info button
}
\description{
getKEGGModal
prepares a helpbutton for to go to a specific site in the documentation
}
\examples{
    x<- getKEGGModal()

}
