% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{mainScatter}
\alias{mainScatter}
\title{mainScatter}
\usage{
mainScatter(data = NULL, x = NULL, y = NULL)
}
\arguments{
\item{data, }{dataframe that has log2FoldChange and log10padj values}

\item{x, }{the name of the x coordinate}

\item{y, }{the name of the y coordinate}
}
\value{
volcano plot
}
\description{
Creates the main scatter plot to be displayed within the main
panel.
}
\examples{
    x <- mainScatter()

}
