% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{runHeatmap}
\alias{runHeatmap}
\title{runHeatmap}
\usage{
runHeatmap(data = NULL, input, num = 1)
}
\arguments{
\item{data, }{a matrixthat includes expression values}

\item{input, }{input varsiables}

\item{num, }{the heatmap number}
}
\value{
heatmap.2 plot
}
\description{
Creates a heatmap based on the user selected parameters within shiny.
}
\examples{
    x <- runHeatmap(mtcars)

}
