% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{startBookmark}
\alias{startBookmark}
\title{startBookmark}
\usage{
startBookmark(session = NULL, input = NULL, Dataset = NULL,
  choicecounter = NULL, buttonValues = NULL, access_token = NULL,
  output = NULL)
}
\arguments{
\item{session, }{sessioin}

\item{input, }{inputs}

\item{Dataset, }{Dataset}

\item{choicecounter, }{choicecounter}

\item{buttonValues, }{buttonValues}

\item{access_token, }{access_token}

\item{output, }{output}
}
\description{
bookmark the selection
}
\examples{
    x <- startBookmark()

}
