% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getDendControls}
\alias{getDendControls}
\title{getDendControls}
\usage{
getDendControls(dendtype = "Row")
}
\arguments{
\item{dendtype, }{Row or Col}
}
\value{
pals
}
\description{
get distance metric parameters
}
\note{
\code{getDendControls}
}
\examples{
    x <- getDendControls()
}
