% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getHeatmapControls}
\alias{getHeatmapControls}
\title{getHeatmapControls}
\usage{
getHeatmapControls(num)
}
\arguments{
\item{num, }{num for repetions of the controls}
}
\value{
HeatmapControls
}
\description{
Generates the left menu to be used for heatmap plots
}
\note{
\code{getHeatmapControls}
}
\examples{
    x <- getHeatmapControls(1)
}
