% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getSizeControls}
\alias{getSizeControls}
\title{getHeatmapControls}
\usage{
getSizeControls(name, h = 500, w = 640)
}
\arguments{
\item{num, }{num for repetions of the controls}
}
\value{
HeatmapControls
}
\description{
Generates the left menu to be used for heatmap plots
}
\note{
\code{getHeatmapControls}
}
\examples{
    x <- getHeatmapControls()
}
