% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{startQCPlots}
\alias{startQCPlots}
\title{startQCPlots}
\usage{
startQCPlots(Dataset = NULL, datasetInput = NULL, cols = NULL,
  conds = NULL, input = NULL, output = NULL)
}
\arguments{
\item{Dataset, }{given data}

\item{datasetInput, }{prepared input data}

\item{cols, }{selected columns}

\item{conds, }{seleced conditions}

\item{input, }{input from ui}

\item{output, }{output}
}
\value{
panel
}
\description{
Gathers the QC plots to be used within the QC panel.
}
\examples{
    startQCPlots()


}
